+++ internal/cli/php_ext.go
@@ -47,7 +47,7 @@ func newPhpExtAddCmd() *cobra.Command {
 			}
 
 			fmt.Printf("Adding extension %q to PHP %s image...\n", ext, version)
-			if err := podman.RebuildFPMImage(version); err != nil {
+			if err := podman.RebuildFPMImage(version, false); err != nil {
 				return err
 			}
 
@@ -89,7 +89,7 @@ func newPhpExtRemoveCmd() *cobra.Command {
 			}
 
 			fmt.Printf("Removing extension %q from PHP %s image...\n", ext, version)
-			if err := podman.RebuildFPMImage(version); err != nil {
+			if err := podman.RebuildFPMImage(version, false); err != nil {
 				return err
 			}